+++ src/main/java/de/samply/email/EmailKeyValuesFactory.java
@@ -1,5 +1,6 @@
 package de.samply.email;
 
+import de.samply.bridgehead.BridgeheadConfiguration;
 import de.samply.db.repository.ProjectBridgeheadRepository;
 import de.samply.db.repository.ProjectRepository;
 import de.samply.db.repository.UserRepository;
@@ -14,21 +15,25 @@ public class EmailKeyValuesFactory {
     private final ProjectBridgeheadRepository projectBridgeheadRepository;
     private final ProjectRepository projectRepository;
     private final UserRepository userRepository;
+    private final BridgeheadConfiguration bridgeheadConfiguration;
 
     public EmailKeyValuesFactory(FrontendService frontendService,
                                  EmailContext emailContext,
                                  ProjectBridgeheadRepository projectBridgeheadRepository,
                                  ProjectRepository projectRepository,
-                                 UserRepository userRepository) {
+                                 UserRepository userRepository,
+                                 BridgeheadConfiguration bridgeheadConfiguration) {
         this.frontendService = frontendService;
         this.emailContext = emailContext;
         this.projectBridgeheadRepository = projectBridgeheadRepository;
         this.projectRepository = projectRepository;
         this.userRepository = userRepository;
+        this.bridgeheadConfiguration = bridgeheadConfiguration;
     }
 
     public EmailKeyValues newInstance() {
-        return new EmailKeyValues(frontendService, emailContext, projectBridgeheadRepository, projectRepository, userRepository);
+        return new EmailKeyValues(frontendService, emailContext, projectBridgeheadRepository, projectRepository,
+                userRepository, bridgeheadConfiguration);
     }
 
 }