+++ CHANGELOG.md
@@ -164,3 +164,4 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Combine first and last name in name 
 - Replace email context variables in email subject
 - Default and Default var in html tags
+- Fetch human readable bridgehead for emails